Please enter the commit message for your changes. Lines starting

+++ udc_inputs.py
@@ -6,7 +6,7 @@ def get_feature_columns(mode):
   feature_columns = []
 
   feature_columns.append(tf.contrib.layers.real_valued_column(
-    column_name="context", dimension=TEXT_FEATURE_SIZE, dtype=tf.int64))
+    column_name="context", dimension=TEXT_FEATURE_SIZE, dtype=tf.float32))
   feature_columns.append(tf.contrib.layers.real_valued_column(
       column_name="context_len", dimension=1, dtype=tf.int64))
   feature_columns.append(tf.contrib.layers.real_valued_column(
@@ -15,13 +15,13 @@ def get_feature_columns(mode):
       column_name="context_nums", dimension=1, dtype=tf.int64))  
 
   feature_columns.append(tf.contrib.layers.real_valued_column(
-      column_name="utterance", dimension=TEXT_FEATURE_SIZE, dtype=tf.int64))
+      column_name="utterance", dimension=TEXT_FEATURE_SIZE, dtype=tf.float32))
   feature_columns.append(tf.contrib.layers.real_valued_column(
       column_name="utterance_len", dimension=1, dtype=tf.int64))
   feature_columns.append(tf.contrib.layers.real_valued_column(
       column_name="utterance_token_len_avg", dimension=1, dtype=tf.float32))
   feature_columns.append(tf.contrib.layers.real_valued_column(
-      column_name="utterance_nums", dimension=1, dtype=tf.int64))  
+      column_name="utterance_nums", dimension=1, dtype=tf.float32))  
   if mode == tf.contrib.learn.ModeKeys.TRAIN:
     # During training we have a label feature
     feature_columns.append(tf.contrib.layers.real_valued_column(
